fix(subsystem): 子系统copy vendors_views 的异常，当 vendors_views 不存在时

+++ generator/subsystem/build/child/build.child.js
@@ -47,10 +47,18 @@ rm(path.resolve(__dirname, '../../dist'), (err) => {
     .then(async () => {
       try {
         await fs.move(path.resolve(__dirname, '../../dist/views'), `${childPath}/views`);
-        await fs.move(
-          path.resolve(__dirname, '../../dist/vendors_views'),
-          `${childPath}/vendors_views`
-        );
+        if (fs.pathExistsSync(path.resolve(__dirname, '../../dist/vendors_views'))) {
+          await fs.move(
+            path.resolve(__dirname, '../../dist/vendors_views'),
+            `${childPath}/vendors_views`
+          );
+        }
+        if (fs.pathExistsSync(path.resolve(__dirname, '../../dist/default_views'))) {
+          await fs.move(
+            path.resolve(__dirname, '../../dist/default_views'),
+            `${childPath}/default_views`
+          );
+        }
       } catch (err) {
         console.error('迁移 views 到子产品目录下出现异常', err);
       }
